Move Perlin and Simplex noises to a core module
