+++ src/core.rs
@@ -1,0 +1,2 @@
+pub mod perlin;
+pub mod simplex;
