+++ src/lib.rs
@@ -17,6 +17,7 @@ extern crate alloc;
 
 pub use crate::noise_fns::*;
 
+pub mod core;
 mod gradient;
 mod math;
 mod noise_fns;
